azure-function/requirements.txt
azure-functions
azure-storage-queue
python-dateutil
